+++ apps/guard/app/(dashboard)/@networks/ItemAddress.tsx
@@ -2,16 +2,9 @@
 
 import React from 'react';
 
-import { Fire, SnowFlake } from '@rosen-bridge/icons';
-import {
-  Amount,
-  Identifier,
-  Stack,
-  SvgIcon,
-  Typography,
-} from '@rosen-bridge/ui-kit';
+import { Amount, Identifier, Stack, Typography } from '@rosen-bridge/ui-kit';
 import { Network as NetworkType } from '@rosen-ui/types';
-import { getAddressUrl, getDecimalString } from '@rosen-ui/utils';
+import { getAddressUrl } from '@rosen-ui/utils';
 
 import { TokenInfoWithAddress } from '@/_types/api';
 
@@ -25,17 +18,11 @@ export const ItemAddress = ({ loading, state, value }: ItemAddressProps) => {
   return (
     <div style={{ width: '100%' }}>
       <Stack
-        spacing={1}
         direction="row"
         align="center"
         justify="start"
         style={{ fontSize: '1.5rem', width: '100%' }}
       >
-        <SvgIcon
-          color={`${state === 'hot' ? 'secondary.dark' : 'tertiary.dark'}`}
-        >
-          {state === 'hot' ? <Fire /> : <SnowFlake />}
-        </SvgIcon>
         <Typography
           color={`${state === 'hot' ? 'secondary.dark' : 'tertiary.dark'}`}
           component="div"
@@ -43,14 +30,10 @@ export const ItemAddress = ({ loading, state, value }: ItemAddressProps) => {
           fontSize="1.5rem"
         >
           <Amount
+            variant={state}
             loading={loading}
-            value={
-              value &&
-              getDecimalString(
-                value.balance.amount.toString(),
-                value.balance.decimals,
-              )
-            }
+            value={value?.balance.amount}
+            decimal={value?.balance.decimals}
             unit={value?.balance.name}
           />
         </Typography>
@@ -58,9 +41,10 @@ export const ItemAddress = ({ loading, state, value }: ItemAddressProps) => {
       <Identifier
         href={getAddressUrl(value?.chain as NetworkType, value?.address)}
         loading={loading}
-        value={value?.address || 'N/A'}
-        copyable={!!value?.address}
-        qrcode={!!value?.address}
+        value={value?.address}
+        fallback="N/A"
+        copyable
+        qrcode
         style={{ minHeight: '30px' }}
       />
     </div>