+++ apps/rosen/src/app/assets/ViewGrid.tsx
@@ -6,6 +6,7 @@ import {
   Chip,
   GridContainer,
   Label,
+  Network,
   Stack,
   Token,
 } from '@rosen-bridge/ui-kit';
@@ -47,12 +48,11 @@ export const ViewGrid = ({
               <Token loading={isLoading} name={item.name} style={{ flex: 1 }} />
               <Chip
                 color="neutral"
-                icon={item.chain?.replace(/(^\w|-\w)/g, (matched) =>
-                  matched.replace('-', '').toUpperCase(),
-                )}
-                label={item.chain}
                 loading={isLoading}
-              />
+                style={{ fontSize: '13px' }}
+              >
+                <Network name={item.chain} />
+              </Chip>
             </Stack>
             <Box mt={1} mb={-1}>
               <Label label="Bridged" dense>