+++ apps/rosen/src/app/events/(list)/EventSidebar.tsx
@@ -84,6 +84,16 @@ const Content = ({ value }: EventSidebarProps) => {
       </Center>
     );
   }
+  if (eventData.status === 'multipleFlows') {
+    return (
+      <Center style={{ minHeight: 'calc(100vh - 304px)' }}>
+        <Typography variant="body1" color="text.secondary" textAlign="center">
+          This event has multiple flows. Click <b>See Details</b> for more
+          information.
+        </Typography>
+      </Center>
+    );
+  }
 
   return <EventDetails value={eventData} showSeeDetailsButton={isTablet} />;
 };