+++ packages/ui-kit/src/components/common/TokensList.tsx
@@ -1,12 +1,10 @@
-import { Fire, SnowFlake } from '@rosen-bridge/icons';
 import { Network, TokenInfoWithColdAmount } from '@rosen-ui/types';
-import { getDecimalString, getTokenUrl } from '@rosen-ui/utils';
+import { getTokenUrl } from '@rosen-ui/utils';
 
 import { Typography } from '../base';
 import { Avatar } from './Avatar';
 import { Amount, Identifier } from './display';
 import { Stack } from './Stack';
-import { SvgIcon } from './SvgIcon';
 import { Text } from './Text';
 
 export interface TokensListProps {
@@ -45,14 +43,12 @@ export const TokensList = ({ chain, tokens, isLoading }: TokensListProps) => (
           }}
         >
           <Avatar
+            loading={isLoading}
+            color="primary"
+            background="primary.light"
+            size={44}
             style={{
               gridRow: 'span 2',
-            }}
-            sx={{
-              color: 'primary.main',
-              bgcolor: 'primary.light',
-              width: 44,
-              height: 44,
             }}
           >
             {token.name?.at(0)}
@@ -62,20 +58,19 @@ export const TokensList = ({ chain, tokens, isLoading }: TokensListProps) => (
           </Text>
           <Stack direction="row" align="center" justify="end" spacing={0.5}>
             {hasCold && (
-              <>
-                <Amount
-                  loading={isLoading}
-                  value={getDecimalString(token.coldAmount, token.decimals)}
-                />
-                <SvgIcon color="tertiary.dark" size="16px">
-                  <SnowFlake />
-                </SvgIcon>
-              </>
+              <Amount
+                variant="cold"
+                reverse
+                loading={isLoading}
+                value={token.coldAmount}
+                decimal={token.decimals}
+              />
             )}
             {!hasCold && (
               <Amount
                 loading={isLoading}
-                value={getDecimalString(token.amount, token.decimals)}
+                value={token.amount}
+                decimal={token.decimals}
               />
             )}
           </Stack>
@@ -95,15 +90,13 @@ export const TokensList = ({ chain, tokens, isLoading }: TokensListProps) => (
           </Typography>
           <Stack direction="row" align="center" justify="end" spacing={0.5}>
             {hasCold && (
-              <>
-                <Amount
-                  loading={isLoading}
-                  value={getDecimalString(token.amount, token.decimals)}
-                />
-                <SvgIcon color="secondary.dark" size="16px">
-                  <Fire />
-                </SvgIcon>
-              </>
+              <Amount
+                variant="hot"
+                reverse
+                loading={isLoading}
+                value={token.amount}
+                decimal={token.decimals}
+              />
             )}
           </Stack>
         </div>